:recycle: move biosample.tasks classes bewteen modules
